Backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.1
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8

